Move HPyErr_ tests into test_hpyerr.
